chore: cleanup and refactor
