+++ packages/cli/src/create-app/createProjectDirectory.cjs
@@ -8,12 +8,14 @@ const fs = require( 'fs/promises' );
  */
 const createProjectDirectory = async ( projectDirPath ) => {
 	try {
+		// Check if the directory exists.
 		await fs.access( projectDirPath );
 	} catch ( error ) {
 		if ( 'ENOENT' !== error.code ) {
 			throw error;
 		}
 
+		// Create the directory if it doesn't exist.
 		await fs.mkdir( projectDirPath, { recursive: true } );
 	}
 };
