+++ packages/cli/src/create-app/printSuccessMessage.cjs
@@ -2,14 +2,14 @@
  * Prints the final instructions for the user.
  *
  * @param {string} projectDirPath - Path to the project directory.
- * @param {boolean} useDefaultPath - Whether using default path or not.
+ * @param {boolean} useDefaultEnv - Whether using default env or not.
  */
-const printInstructions = ( projectDirPath, useDefaultPath ) => {
+const printSuccessMessage = ( projectDirPath, useDefaultEnv ) => {
 	console.log( '' );
 	console.log( `Your project has been scaffolded at: ${ projectDirPath }.` );
 	console.log( '' );
 
-	if ( useDefaultPath ) {
+	if ( useDefaultEnv ) {
 		console.log(
 			'For setting up environment variables, please refer to the documentation at: https://github.com/rtCamp/snapwp/blob/b7c0472d95be624244ad2a5d01d4bcdaa29e91f3/packages/cli/README.md'
 		);
@@ -25,5 +25,5 @@ const printInstructions = ( projectDirPath, useDefaultPath ) => {
 };
 
 module.exports = {
-	printInstructions,
+	printSuccessMessage,
 };
