+++ packages/cli/src/create-app/setupNpmrc.cjs
@@ -1,6 +1,8 @@
 const path = require( 'path' );
 const fs = require( 'fs/promises' );
-const { NPMRC_CONTENT, REGISTRY_URL } = require( '../constant.cjs' );
+
+const REGISTRY_URL = 'http://localhost:4873';
+const NPMRC_CONTENT = `@snapwp:registry=${ REGISTRY_URL }`;
 
 /**
  * Creates .npmrc file for proxy registry if needed.
@@ -12,10 +14,12 @@ const { NPMRC_CONTENT, REGISTRY_URL } = require( '../constant.cjs' );
 const setupNpmrc = async ( projectDirPath, useProxy ) => {
 	if ( useProxy ) {
 		console.log( 'Found --proxy flag, generating `.npmrc` file.' );
+
 		await fs.writeFile(
 			path.join( projectDirPath, '.npmrc' ),
 			NPMRC_CONTENT
 		);
+
 		console.log(
 			`\`.npmrc\` file generated successfully. Please make sure the proxy registry is running on ${ REGISTRY_URL }`
 		);
