+++ packages/cli/src/utils/prompt.cjs
@@ -13,6 +13,7 @@ const prompt = ( query, defaultValue = '' ) => {
 		input: process.stdin,
 		output: process.stdout,
 	} );
+
 	return new Promise( ( resolve ) => {
 		rl.question(
 			// Append default value to the query.
